+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - SDK GithubActions generator
 - SDK Readme generator
+- CORS Filter
 
 ## [0.0.22] 2021-04-02
 ### Added